Turned into modules for Android. Compiling.
Remaining: Tune builder to integrate modulus only.
Run Jitbrains on the result to source externally the package.
