+++ AndroidApp/.idea/modules.xml
@@ -4,6 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/AndroidApp.iml" filepath="$PROJECT_DIR$/AndroidApp.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/qr-authserver/qr-authserver.iml" filepath="$PROJECT_DIR$/qr-authserver/qr-authserver.iml" />
     </modules>
   </component>
 </project>
