+++ AndroidApp/app/src/main/java/fr/jserv/androidapp/LoginActivity.java
@@ -1,225 +1,28 @@
 package fr.jserv.androidapp;
-
-
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.content.Intent;
-import android.util.Log;
-import android.view.View;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import com.github.nkzawa.socketio.client.IO;
-import com.github.nkzawa.socketio.client.Socket;
-import com.github.nkzawa.emitter.Emitter;
+import fr.jserv.qr_authserver.QRSignIn;
 
-import org.apache.http.NameValuePair;
-import org.apache.http.client.utils.URLEncodedUtils;
+public class LoginActivity extends QRSignIn {
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-
-public class LoginActivity extends AppCompatActivity {
-
-    private Socket mSocket;
-    private String serverUrl;
-    private TextView tvMain;
-    private EditText tvName, tvSurname;
-
-    public LoginActivity getActivity() {
-        return this;
-    }
-
-    private Emitter.Listener onChallenge = new Emitter.Listener() {
-        @Override
-        public void call(final Object... args) {
-            getActivity().runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    String response = answerChallenge(args);
-                    mSocket.emit("uuidresponse", response);
-                }
-            });
-        }
-    };
-
-    private Emitter.Listener onSuccessSignIn = new Emitter.Listener() {
-        @Override
-        public void call(final Object... args) {
-            getActivity().runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    addMessage("Successful Sign-In !");
-                    setCredsVisibility(View.INVISIBLE);
-                }
-            });
-        }
-    };
-    private Emitter.Listener onSuccessLogIn = new Emitter.Listener() {
-        @Override
-        public void call(final Object... args) {
-            getActivity().runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    addMessage("Successful Log In !");
-                    setCredsVisibility(View.VISIBLE);
-                    JSONObject data = (JSONObject) args[0];
-                    try {
-                        tvName.setText(data.getString("name"));
-                        tvName.setText(data.getString("Surname"));
-                    } catch (JSONException e) {
-                        e.printStackTrace();
-                    }
-                }
-            });
-        }
-    };
-    private Emitter.Listener onSuccessBind = new Emitter.Listener() {
-        @Override
-        public void call(final Object... args) {
-            getActivity().runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    addMessage("Successful Bind !");
-                }
-            });
-        }
-    };
-    private Emitter.Listener onUpdate = new Emitter.Listener() {
-        @Override
-        public void call(final Object... args) {
-            getActivity().runOnUiThread(new Runnable() {
-                @Override
-                public void run() {
-                    JSONObject data = (JSONObject) args[0];
-                    Log.i("mSocket", "Got update request: " + data.toString());
-
-                    String message;
-                    try {
-                        switch(data.getString("id")){
-                            case "txtName":
-                                message = data.getString("value");
-                                Log.d("mSocket", "id field contains:" + message);
-                                tvName.setText(message);
-                                break;
-                            case "txtSurname":
-                                message = data.getString("value");
-                                Log.d("mSocket", "id field contains:" + message);
-                                tvSurname.setText(message);
-                                break;
-                            default:
-                                throw new Exception("No component matching form.");
-                        }
-
-                    } catch (JSONException e) {
-                        e.printStackTrace();
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                }
-            });
-        }
-    };
-
-    private String answerChallenge(Object... args) {
+    protected String answerChallenge(Object... args) {
         return "42";
     }
 
-    private void addMessage(String message){
-        tvMain.setText(message);
-    }
+    protected void onCreate(Bundle savedInstanceState){
+        this.tvMain = (TextView) findViewById(R.id.tvMain);
+        this.lblName = (TextView) findViewById(R.id.lblName);
+        this.lblSurname = (TextView) findViewById(R.id.lblSurname);
+        this.tvName = (EditText) findViewById(R.id.tvName);
+        this.tvSurname = (EditText) findViewById(R.id.tvSurname);
 
-    private void setCredsVisibility(int visibility){
-        tvName.setVisibility(visibility);
-        tvSurname.setVisibility(visibility);
-        ((TextView) findViewById(R.id.lblName)).setText(visibility);
-        ((TextView) findViewById(R.id.lblSurname)).setText(visibility);
-    }
-
-
-
-    private void update(String id, String text){
-        JSONObject data = new JSONObject();
-        try {
-            data.put("id", id);
-            data.put("value", text);
-            Log.d("mSocket", "Attempting to send Object: " + data.toString());
-
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-        mSocket.emit("broadcast", data);
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
-        String psk;
         setContentView(R.layout.activity_login);
-        tvMain = (TextView) findViewById(R.id.tvMain);
-        tvName = (EditText) findViewById(R.id.tvName);
-        tvSurname = (EditText) findViewById(R.id.tvSurname);
-        tvMain.setText("Pending...");
-        Intent intent = getIntent();
-        serverUrl = intent.getStringExtra(MainActivity.EXTRA_MESSAGE).toUpperCase();
-        URI url;
-        List<NameValuePair> params = null;
-        try {
-            url = new URI(serverUrl);
-
-            params = URLEncodedUtils.parse(url, "UTF-8");
-            if(params.get(0).getName().toUpperCase().equals("ID")) {
-                psk = params.get(0).getValue().toLowerCase();
-            }
-            else {
-                throw new URISyntaxException(serverUrl, "id not present amongst name:" + params.get(0).getName() + " and value : " + params.get(0).getValue());
-            }
-            mSocket = IO.socket(url.getScheme().toLowerCase()+"://"+url.getRawAuthority().toLowerCase());
-            mSocket.connect();
-            Log.i("mSocket", "Connecting to URL " + url.getScheme().toLowerCase()+"://"+url.getRawAuthority().toLowerCase() + " successful.");
-
-            mSocket.emit("authid", psk);
-            Log.d("mSocket", "Sending authid with PSK : " + psk);
-
-            mSocket.on("uuidquery", onChallenge);
-            mSocket.on("Successful Signin", onSuccessSignIn);
-            mSocket.on("Successful Bind", onSuccessBind);
-            mSocket.on("Successful Login", onSuccessLogIn);
-            mSocket.on("update", onUpdate);
-        } catch (URISyntaxException e) {
-            e.printStackTrace();
-        }
-
-        tvName.setOnFocusChangeListener(new View.OnFocusChangeListener() {
-            @Override
-            public void onFocusChange(View v, boolean hasFocus) {
-                if (!hasFocus) update("txtName", tvName.getText().toString());
-            }
-        });
-        tvSurname.setOnFocusChangeListener(new View.OnFocusChangeListener() {
-            @Override
-            public void onFocusChange(View v, boolean hasFocus) {
-                if (!hasFocus) update("txtSurname", tvSurname.getText().toString());
-            }
-        });
-
     }
 
 
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
 
-        mSocket.disconnect();
-        mSocket.off("uuidquery", onChallenge);
-        mSocket.off("Successful Signin", onSuccessSignIn);
-        mSocket.off("Successful Bind", onSuccessBind);
-        mSocket.off("Successful Login", onSuccessLogIn);
-        mSocket.off("update", onUpdate);
-    }
+
 }
