+++ AndroidApp/qr-authserver/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">QR-Auth Server</string>
+
+</resources>
